requirements.txt
asgiref==3.6.0
autopep8==2.0.2
backports.zoneinfo==0.2.1
Django==4.2
djangorestframework==3.14.0
numpy==1.24.3
pandas==2.0.1
psycopg2==2.9.6
psycopg2-binary==2.9.6
pycodestyle==2.10.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
tzdata==2023.3
